New version v1.1.0

+++ src/components/AdblockCase/selectByUser.js
@@ -1,9 +1,12 @@
-import Axios from 'axios';
 import { useEffect, useState } from 'react';
+import Axios from 'axios';
+import $ from 'jquery';
 
 function SelectByUser() {
 
     const [selectState, setSelectState] = useState(false)
+    const [state, setState] = useState(null)
+    const [status, setStatus] = useState(null)
 
     useEffect(() => {
         Axios.get(`https://restcountries.com/v3.1/all`)
@@ -46,31 +49,56 @@ function SelectByUser() {
                     setSelectState(false);
                 }).catch(erro => { console.log(erro) })
 
-            document.querySelector('.bttnn').innerHTML = "=>"
+            document.querySelector('.bttnn').innerHTML = "=> Check"
         };
 
     }, [selectState])
 
 
     const save = () => {
-        localStorage.clear();
-        var num = document.querySelector('.elsex').selectedIndex - 1
-        localStorage.setItem('country', document.querySelectorAll('.elsex option')[num].name)
-        localStorage.setItem('city', document.querySelector('.ifsa').value)
-        document.querySelector('.bttnn').innerHTML = "Redy!";
-        document.querySelector('#gohome').href = "/";
+        setState(false);
+        document.querySelector('.stt').innerHTML = "";
+        if (document.querySelector('.elsex').value === 'Select Country' || document.querySelector('.ifsa').value === 'Select Country' || document.querySelector('.ifsa').value === '') {
+            setStatus(false);
+            setState(true);
+            document.querySelector('.stt').innerHTML = "Please fill all feild, then submit."
+        } else {
+            var num = document.querySelector('.elsex').selectedIndex - 1;
+            var cc = document.querySelectorAll('.elsex option')[num].name;
+            var ss = document.querySelector('.ifsa').value;
+            $(
+                $.getJSON(`https://muslimsalat.com/${cc}/${ss}.json?key=9233c34903ef6aa6fd59a97cedac8226&jsoncallback=?`, function (data) {
+                    if (data.status_code === 0) {
+                        setStatus(false);
+                        setState(true);
+                        document.querySelector('.stt').innerHTML = "We're really sorry, but we dont support your location yet. Try to chose a big city in your county."
+
+                    } else {
+                        localStorage.clear();
+                        setStatus(true);
+                        setState(true);
+                        localStorage.setItem('country', cc)
+                        localStorage.setItem('city', ss)
+                        document.querySelector('.bttnn').innerHTML = "Redy!";
+                        document.querySelector('.stt').innerHTML = "All done! Enjoy the app."
+                        window.location.replace("/");
+                    }
+                })
+            )
+
+        }
     }
 
 
     return (
         <>
-            <div className="lg:w-1/2 w-full h-[26em] md:h-auto shadow-xl flex md:flex-row flex-col justify-between items-center p-4 rounded-lg bg-slate-100 dark:bg-gray-800 dark:text-slate-100 text-gray-800">
-                <div className="flex flex-col h-full justify-evenly items-center">
-                    <div className="flex flex-row items-end py-6 space-x-2">
-                        <header className="text-2xl font-bold">Fill your Location.</header>
-                        <i>its quickly.</i>
+            <div className="lg:w-1/2 w-full md:h-auto shadow-xl flex flex-col justify-between items-stretch p-4 rounded-lg bg-slate-100 dark:bg-gray-800 dark:text-slate-100 text-gray-800">
+                <div className="flex flex-col h-full w-full justify-evenly items-center">
+                    <div className="flex flex-row items-end w-[87%] py-6 space-x-2">
+                        <header className="text-2xl lg:text-4xl font-bold">Fill your Location.</header>
+                        <i>easy way!</i>
                     </div>
-                    <div className="flex flex-col space-y-4 p-2">
+                    <div className="flex flex-col w-[90%] space-y-4 p-2">
                         <div>
                             <select class="select select-info w-full elsex">
                                 <option disabled selected>Select Country</option>
@@ -83,7 +111,22 @@ function SelectByUser() {
                         </div>
                     </div>
                 </div>
-                <a id='gohome' href='#'><button onClick={save} className="bttnn btn btn-info md:mt-0 mt-6 sm:h-[100%] md:w-[8em] h-[4em] w-[4em] text-lg font-bold">{"->"}</button></a>
+                <div className='w-full mt-[1em] flex flex-col-reverse items-end justify-start'>
+                    <a id='gohome' className='w-full flex justify-center' href='#__'><button onClick={save} className="bttnn btn btn-info sm:mt-0 mt-6 w-[90%] h-[4em] text-lg font-bold">{"=> Check"}</button></a>
+                    <div className='w-[90%] mb-[1em]'>
+                        {
+                            state === false ?
+                                <div className="loader py-[2em]">
+                                    <div className="outer"></div>
+                                    <div className="middle"></div>
+                                    <div className="inner"></div>
+                                </div>
+                                :
+                                <div className="stt text-xl font-extrabold " >
+                                </div>
+                        }
+                    </div>
+                </div>
             </div>
         </>
     );
